Eagle level completed
I added the Eagle level assignment. Not sure if this is really the most effective
way to do this, so more than happy for better ideas on how to implement, but this
should work.

Also replaced the test for stand for player if they bust with the should_receive
method from rspec  - very cool, thx for pointing that out jwo!